Subindo o repositorio do projeto no github

+++ database/migrations/2022_01_11_161944_criar_tabela_contratos.php
@@ -20,7 +20,7 @@ class CriarTabelaContratos extends Migration
                 $table->string('objeto');
                 $table->string('idEmpresa');
                 $table->string('idSecretaria');
-                $table->float('valor');
+                $table->decimal('valor');
                 $table->string('vencimento');
             }
         );
